requirements.txt
autopep8==1.5.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
openpyxl==3.0.3
pandas==1.0.3
pkg-resources==0.0.0
protobuf==3.11.3
protobuf3-to-dict==0.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
Shapely==1.7.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
